+++ src/api/v1/ApiRouter.cpp
@@ -102,21 +102,12 @@ void xmrig::ApiRouter::getHashrate(rapidjson::Value &reply, rapidjson::Document 
 void xmrig::ApiRouter::getMiner(rapidjson::Value &reply, rapidjson::Document &doc) const
 {
     auto &allocator = doc.GetAllocator();
-    auto &stats = static_cast<Controller *>(m_base)->statsData();
 
     reply.AddMember("version",      APP_VERSION, allocator);
     reply.AddMember("kind",         APP_KIND, allocator);
     reply.AddMember("algo",         "invalid", allocator);
     reply.AddMember("mode",         rapidjson::StringRef(m_base->config()->modeName()), allocator);
     reply.AddMember("ua",           Platform::userAgent().toJSON(), allocator);
-    reply.AddMember("donate_level", m_base->config()->pools().donateLevel(), allocator);
-
-    if (stats.hashes && stats.donateHashes) {
-        reply.AddMember("donated", normalize((double) stats.donateHashes / stats.hashes * 100.0), allocator);
-    }
-    else {
-        reply.AddMember("donated", 0.0, allocator);
-    }
 }
 
 
@@ -205,7 +196,6 @@ void xmrig::ApiRouter::getResults(rapidjson::Value &reply, rapidjson::Document &
     results.AddMember("avg_time",      stats.avgTime(), allocator);
     results.AddMember("latency",       stats.avgLatency(), allocator);
     results.AddMember("hashes_total",  stats.hashes, allocator);
-    results.AddMember("hashes_donate", stats.donateHashes, allocator);
 
     rapidjson::Value best(rapidjson::kArrayType);
     for (uint64_t i : stats.topDiff) {