+++ src/base/kernel/config/BaseTransform.cpp
@@ -242,7 +242,6 @@ void xmrig::BaseTransform::transform(rapidjson::Document &doc, int key, const ch
     case IConfig::RetryPauseKey:    /* --retry-pause */
     case IConfig::PrintTimeKey:     /* --print-time */
     case IConfig::HttpPort:         /* --http-port */
-    case IConfig::DonateLevelKey:   /* --donate-level */
     case IConfig::DaemonPollKey:    /* --daemon-poll-interval */
     case IConfig::DaemonJobTimeoutKey: /* --daemon-job-timeout */
     case IConfig::DnsTtlKey:        /* --dns-ttl */
@@ -341,12 +340,6 @@ void xmrig::BaseTransform::transformUint64(rapidjson::Document &doc, int key, ui
 
     case IConfig::RetryPauseKey: /* --retry-pause */
         return set(doc, Pools::kRetryPause, arg);
-
-    case IConfig::DonateLevelKey: /* --donate-level */
-        return set(doc, Pools::kDonateLevel, arg);
-
-    case IConfig::ProxyDonateKey: /* --donate-over-proxy */
-        return set(doc, Pools::kDonateOverProxy, arg);
 
     case IConfig::HttpPort: /* --http-port */
         m_http = true;