+++ src/base/net/stratum/Pools.cpp
@@ -29,7 +29,6 @@
 #include "base/kernel/interfaces/IJsonReader.h"
 #include "base/net/stratum/strategies/FailoverStrategy.h"
 #include "base/net/stratum/strategies/SinglePoolStrategy.h"
-#include "donate.h"
 
 
 #ifdef XMRIG_FEATURE_BENCHMARK
@@ -40,8 +39,6 @@
 namespace xmrig {
 
 
-const char *Pools::kDonateLevel     = "donate-level";
-const char *Pools::kDonateOverProxy = "donate-over-proxy";
 const char *Pools::kPools           = "pools";
 const char *Pools::kRetries         = "retries";
 const char *Pools::kRetryPause      = "retry-pause";
@@ -50,8 +47,7 @@ const char *Pools::kRetryPause      = "retry-pause";
 } // namespace xmrig
 
 
-xmrig::Pools::Pools() :
-    m_donateLevel(kDefaultDonateLevel)
+xmrig::Pools::Pools()
 {
 #   ifdef XMRIG_PROXY_PROJECT
     m_retries    = 2;
@@ -67,16 +63,6 @@ bool xmrig::Pools::isEqual(const Pools &other) const
     }
 
     return std::equal(m_data.begin(), m_data.end(), other.m_data.begin());
-}
-
-
-int xmrig::Pools::donateLevel() const
-{
-#   ifdef XMRIG_FEATURE_BENCHMARK
-    return benchSize() || (m_benchmark && !m_benchmark->id().isEmpty()) ? 0 : m_donateLevel;
-#   else
-    return m_donateLevel;
-#   endif
 }
 
 
@@ -158,8 +144,6 @@ void xmrig::Pools::load(const IJsonReader &reader)
         }
     }
 
-    setDonateLevel(reader.getInt(kDonateLevel, kDefaultDonateLevel));
-    setProxyDonate(reader.getInt(kDonateOverProxy, PROXY_DONATE_AUTO));
     setRetries(reader.getInt(kRetries));
     setRetryPause(reader.getInt(kRetryPause));
 }
@@ -207,33 +191,9 @@ void xmrig::Pools::toJSON(rapidjson::Value &out, rapidjson::Document &doc) const
     }
 #   endif
 
-    doc.AddMember(StringRef(kDonateLevel),      m_donateLevel, allocator);
-    doc.AddMember(StringRef(kDonateOverProxy),  m_proxyDonate, allocator);
     out.AddMember(StringRef(kPools),            toJSON(doc), allocator);
     doc.AddMember(StringRef(kRetries),          retries(), allocator);
     doc.AddMember(StringRef(kRetryPause),       retryPause(), allocator);
-}
-
-
-void xmrig::Pools::setDonateLevel(int level)
-{
-    if (level >= kMinimumDonateLevel && level <= 99) {
-        m_donateLevel = level;
-    }
-}
-
-
-void xmrig::Pools::setProxyDonate(int value)
-{
-    switch (value) {
-    case PROXY_DONATE_NONE:
-    case PROXY_DONATE_AUTO:
-    case PROXY_DONATE_ALWAYS:
-        m_proxyDonate = static_cast<ProxyDonate>(value);
-
-    default:
-        break;
-    }
 }