+++ src/core/config/Config.h
@@ -57,10 +57,9 @@ public:
     void getJSON(rapidjson::Document &doc) const override;
     void toggleVerbose();
 
-    inline bool hasAlgoExt() const                 { return isDonateOverProxy() ? m_algoExt : true; }
+    inline bool hasAlgoExt() const                 { return m_algoExt; }
     inline bool isCustomDiffStats() const          { return m_customDiffStats; }
     inline bool isDebug() const                    { return m_debug; }
-    inline bool isDonateOverProxy() const          { return m_pools.donateLevel() == 0 || m_mode == SIMPLE_MODE; }
     inline bool isShouldSave() const               { return m_upgrade && isAutoSave(); }
     inline const BindHosts &bind() const           { return m_bind; }
     inline const String &accessLog() const         { return m_accessLog; }