+++ src/proxy/Proxy.cpp
@@ -48,7 +48,6 @@
 #include "proxy/Miners.h"
 #include "proxy/ProxyDebug.h"
 #include "proxy/Server.h"
-#include "proxy/splitters/donate/DonateSplitter.h"
 #include "proxy/splitters/extra_nonce/ExtraNonceSplitter.h"
 #include "proxy/splitters/nicehash/NonceSplitter.h"
 #include "proxy/splitters/simple/SimpleSplitter.h"
@@ -90,7 +89,6 @@ xmrig::Proxy::Proxy(Controller *controller) :
     }
 
     m_splitter  = splitter;
-    m_donate    = new DonateSplitter(controller);
     m_stats     = new Stats(controller);
     m_shareLog  = new ShareLog(controller, m_stats);
     m_accessLog = new AccessLog(controller);
@@ -107,21 +105,18 @@ xmrig::Proxy::Proxy(Controller *controller) :
     Events::subscribe(IEvent::ConnectionType, m_stats);
 
     Events::subscribe(IEvent::CloseType, m_miners);
-    Events::subscribe(IEvent::CloseType, m_donate);
     Events::subscribe(IEvent::CloseType, splitter);
     Events::subscribe(IEvent::CloseType, m_stats);
     Events::subscribe(IEvent::CloseType, m_accessLog);
     Events::subscribe(IEvent::CloseType, m_workers);
 
     Events::subscribe(IEvent::LoginType, m_login);
-    Events::subscribe(IEvent::LoginType, m_donate);
     Events::subscribe(IEvent::LoginType, &m_customDiff);
     Events::subscribe(IEvent::LoginType, splitter);
     Events::subscribe(IEvent::LoginType, m_stats);
     Events::subscribe(IEvent::LoginType, m_accessLog);
     Events::subscribe(IEvent::LoginType, m_workers);
 
-    Events::subscribe(IEvent::SubmitType, m_donate);
     Events::subscribe(IEvent::SubmitType, splitter);
     Events::subscribe(IEvent::SubmitType, m_stats);
     Events::subscribe(IEvent::SubmitType, m_workers);
@@ -150,7 +145,6 @@ xmrig::Proxy::~Proxy()
     delete m_api;
 #   endif
 
-    delete m_donate;
     delete m_login;
     delete m_miners;
     delete m_splitter;