+++ src/proxy/Stats.cpp
@@ -123,10 +123,6 @@ void xmrig::Stats::accept(const AcceptEvent *event)
     m_data.accepted++;
     m_data.hashes += event->result.diff;
 
-    if (event->isDonate()) {
-        m_data.donateHashes += event->result.diff;
-    }
-
     Counters::accepted++;
 
     const size_t ln = m_data.topDiff.size() - 1;
@@ -139,11 +135,7 @@ void xmrig::Stats::accept(const AcceptEvent *event)
 }
 
 
-void xmrig::Stats::reject(const AcceptEvent *event)
+void xmrig::Stats::reject(const AcceptEvent *)
 {
-    if (event->isDonate()) {
-        return;
-    }
-
     m_data.rejected++;
 }