+++ src/proxy/log/ShareLog.cpp
@@ -76,7 +76,7 @@ void xmrig::ShareLog::onRejectedEvent(IEvent *event)
 
 void xmrig::ShareLog::accept(const AcceptEvent *event)
 {
-    if (!m_controller->config()->isVerbose() || event->isDonate() || event->isCustomDiff()) {
+    if (!m_controller->config()->isVerbose() || event->isCustomDiff()) {
         return;
     }
 
@@ -87,10 +87,6 @@ void xmrig::ShareLog::accept(const AcceptEvent *event)
 
 void xmrig::ShareLog::reject(const AcceptEvent *event)
 {
-    if (event->isDonate()) {
-        return;
-    }
-
     LOG_INFO("%s " CYAN("%04u ") RED_BOLD("rejected") " (%" PRId64 "/%" PRId64 "+%" PRId64 ") diff " WHITE_BOLD("%" PRIu64) " ip " WHITE_BOLD("%s") " " RED("\"%s\"") " " BLACK_BOLD("(%" PRIu64 " ms)"),
              Tags::proxy(), event->mapperId(), m_stats->data().accepted, m_stats->data().rejected, m_stats->data().invalid, event->result.diff, event->ip(), event->error(), event->result.elapsed);
 }