+++ src/proxy/splitters/extra_nonce/ExtraNonceMapper.cpp
@@ -29,12 +29,12 @@
 #include "proxy/splitters/extra_nonce/ExtraNonceMapper.h"
 #include "base/io/log/Log.h"
 #include "base/io/log/Tags.h"
+#include "base/kernel/interfaces/IStrategy.h"
 #include "base/net/stratum/Client.h"
 #include "base/net/stratum/Pools.h"
 #include "core/config/Config.h"
 #include "core/Controller.h"
 #include "net/JobResult.h"
-#include "net/strategies/DonateStrategy.h"
 #include "proxy/Error.h"
 #include "proxy/events/AcceptEvent.h"
 #include "proxy/events/SubmitEvent.h"
@@ -47,10 +47,6 @@ xmrig::ExtraNonceMapper::ExtraNonceMapper(size_t, Controller *controller) :
 {
     m_storage  = new ExtraNonceStorage();
     m_strategy = controller->config()->pools().createStrategy(this);
-
-    if (controller->config()->pools().donateLevel() > 0) {
-        m_donate = new DonateStrategy(controller, this);
-    }
 }
 
 
@@ -59,7 +55,6 @@ xmrig::ExtraNonceMapper::~ExtraNonceMapper()
     delete m_pending;
     delete m_strategy;
     delete m_storage;
-    delete m_donate;
 }
 
 
@@ -133,24 +128,13 @@ void xmrig::ExtraNonceMapper::submit(SubmitEvent *event)
     JobResult req = event->request;
     req.diff = m_storage->job().diff();
 
-    IStrategy *strategy = m_donate && m_donate->isActive() ? m_donate : m_strategy;
-
-    m_results[strategy->submit(req)] = SubmitCtx(req.id, event->miner()->id());
+    m_results[m_strategy->submit(req)] = SubmitCtx(req.id, event->miner()->id());
 }
 
 
 void xmrig::ExtraNonceMapper::tick(uint64_t, uint64_t now)
 {
     m_strategy->tick(now);
-
-    if (m_donate) {
-        m_donate->tick(now);
-
-        if (m_donate->isActive() && m_donate->hasPendingJob() && m_donate->reschedule()) {
-            const auto &pending = m_donate->pending();
-            setJob(pending.host.data(), pending.port, pending.job);
-        }
-    }
 }
 
 
@@ -197,15 +181,6 @@ void xmrig::ExtraNonceMapper::onActive(IStrategy *strategy, IClient *client)
 
 void xmrig::ExtraNonceMapper::onJob(IStrategy *, IClient *client, const Job &job, const rapidjson::Value &)
 {
-    if (m_donate) {
-        if (m_donate->isActive() && client->id() != -1 && !m_donate->reschedule()) {
-            m_donate->save(client, job);
-            return;
-        }
-
-        m_donate->setAlgo(job.algorithm());
-    }
-
     setJob(client->pool().host(), client->pool().port(), job);
 }
 
@@ -226,11 +201,11 @@ void xmrig::ExtraNonceMapper::onPause(IStrategy *)
 }
 
 
-void xmrig::ExtraNonceMapper::onResultAccepted(IStrategy *, IClient *client, const SubmitResult &result, const char *error)
+void xmrig::ExtraNonceMapper::onResultAccepted(IStrategy *, IClient *, const SubmitResult &result, const char *error)
 {
     const SubmitCtx ctx = submitCtx(result.seq);
 
-    AcceptEvent::start(0, ctx.miner, result, client->id() == -1, false, error);
+    AcceptEvent::start(0, ctx.miner, result, false, error);
 
     if (!ctx.miner) {
         return;
@@ -273,10 +248,6 @@ void xmrig::ExtraNonceMapper::connect()
 {
     m_suspended = 0;
     m_strategy->connect();
-
-    if (m_donate) {
-        m_donate->connect();
-    }
 }
 
 
@@ -297,8 +268,4 @@ void xmrig::ExtraNonceMapper::suspend()
     m_storage->setActive(false);
     m_storage->reset();
     m_strategy->stop();
-
-    if (m_donate) {
-        m_donate->stop();
-    }
 }